+++ torch/csrc/jit/function_schema_parser.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <ATen/core/function_schema.h>
+#include <torch/csrc/WindowsTorchApiMacro.h>
+#include <string>
+
+namespace torch {
+namespace jit {
+
+TORCH_API ::c10::FunctionSchema parseSchema(const std::string& schema);
+
+} // namespace jit
+} // namespace torch
