+++ torch/csrc/jit/script/schema_type_parser.h
@@ -1,6 +1,5 @@
-#include <ATen/ATen.h>
 #include <ATen/core/jit_type.h>
-#include <torch/csrc/jit/alias_info.h>
+#include <ATen/core/alias_info.h>
 #include <torch/csrc/jit/script/lexer.h>
 
 namespace torch {
@@ -8,12 +7,12 @@ namespace jit {
 namespace script {
 
 using TypePtr = c10::TypePtr;
-using TypeAndAlias = std::pair<TypePtr, c10::optional<AliasInfo>>;
+using TypeAndAlias = std::pair<TypePtr, c10::optional<c10::AliasInfo>>;
 
 struct SchemaTypeParser {
   TypeAndAlias parseBaseType();
-  c10::optional<AliasInfo> parseAliasAnnotation();
-  std::pair<TypePtr, c10::optional<AliasInfo>> parseType();
+  c10::optional<c10::AliasInfo> parseAliasAnnotation();
+  std::pair<TypePtr, c10::optional<c10::AliasInfo>> parseType();
   c10::optional<at::ScalarType> parseTensorDType(const std::string& dtype);
   TypePtr parseRefinedTensor();
